@@ -1,40 +1,48 @@
 apply plugin: 'com.android.application'
+apply plugin: 'com.google.gms.google-services'
+
+
+repositories {
+    maven { url 'https://jitpack.io' }
+}
 
 dependencies {
     api fileTree(include: ['*.jar'], dir: 'libs')
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
-    implementation 'com.google.android.gms:play-services-plus:16.0.0'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
+    implementation 'com.google.android.gms:play-services-plus:17.0.0'
     testImplementation 'junit:junit:4.12'
-    testImplementation 'org.mockito:mockito-core:1.10.19'
-    api 'com.android.support:appcompat-v7:26.1.0'
-    api 'com.android.support:support-v4:26.1.0'
-    api 'com.android.support:design:26.1.0'
-    api 'joda-time:joda-time:2.3'
-    api 'com.google.android.gms:play-services-location:9.2.0'
+    testImplementation 'org.mockito:mockito-core:2.15.0'
+    api 'androidx.appcompat:appcompat:1.0.2'
+    api 'androidx.legacy:legacy-support-v4:1.0.0'
+    api 'com.google.android.material:material:1.0.0'
+    api 'joda-time:joda-time:2.9.9'
+    api 'com.google.android.gms:play-services-location:17.0.0'
     api 'com.j256.ormlite:ormlite-core:5.1'
     api 'com.j256.ormlite:ormlite-android:5.1'
     api 'org.reflections:reflections:0.9.10'
-    api 'com.google.firebase:firebase-core:16.0.4'
-    api 'com.google.firebase:firebase-messaging:17.3.3'
+    api 'com.google.firebase:firebase-core:17.0.0'
+    api 'com.google.firebase:firebase-messaging:19.0.0'
     api 'org.bouncycastle:bcprov-jdk15on:1.60'
     api 'org.bouncycastle:bcpkix-jdk15on:1.60'
-    implementation 'com.google.code.gson:gson:2.3.1'
-    api 'com.android.support.constraint:constraint-layout:1.1.0'
+    api 'com.google.code.gson:gson:2.8.5'
+    api 'androidx.constraintlayout:constraintlayout:1.1.3'
     api 'org.springframework.android:spring-android-rest-template:1.0.1.RELEASE'
-    api 'com.fasterxml.jackson.core:jackson-databind:2.3.2'
-    api 'org.apache.httpcomponents:httpclient:4.5.5'
+    api 'com.fasterxml.jackson.core:jackson-databind:2.9.6'
+    api 'org.apache.httpcomponents:httpclient:4.5.6'
     api 'com.android.volley:volley:1.1.1'
     api 'com.loopj.android:android-async-http:1.4.9'
+    api 'com.github.PhilJay:MPAndroidChart:v3.1.0'
+    api 'com.google.android.material:material:1.0.0'
 }
 
 android {
-    compileSdkVersion 26
+    compileSdkVersion 29
     buildToolsVersion '28.0.3'
 
     defaultConfig {
         applicationId "com.sonu.diary"
         minSdkVersion 24
-        targetSdkVersion 26
+        targetSdkVersion 29
         versionCode 3
         versionName "1.0"
     }
