@@ -4,7 +4,9 @@ import android.app.DatePickerDialog;
 import android.app.Dialog;
 import android.content.Context;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatDialogFragment;
+
+
+import androidx.appcompat.app.AppCompatDialogFragment;
 
 import com.sonu.diary.activities.AbstractActivity;
 
