@@ -1,7 +1,9 @@
 package com.sonu.diary.activities;
 
-import android.support.v7.app.AppCompatActivity;
+
 import android.os.Bundle;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.sonu.diary.R;
 
