@@ -1,6 +1,6 @@
 package com.sonu.diary.activities;
 
-import android.support.v7.app.AppCompatActivity;
+
 import android.os.Bundle;
 import android.view.View;
 import android.widget.EditText;
