@@ -4,7 +4,7 @@ import com.j256.ormlite.field.DatabaseField;
 import com.j256.ormlite.table.DatabaseTable;
 
 @DatabaseTable(tableName = "payment_modes")
-public class PaymentMode {
+public class PaymentMode extends AbstractDomain {
 
     @DatabaseField
     private String description;
@@ -15,5 +15,10 @@ public class PaymentMode {
 
     public void setDescription(String description) {
         this.description = description;
+    }
+
+    @Override
+    public String getKey() {
+        return description;
     }
 }
