@@ -1,11 +1,11 @@
 package com.sonu.diary.services;
 
+import com.sonu.diary.remote.SyncService;
+
 import org.junit.Test;
 
 import java.util.LinkedList;
 import java.util.List;
-
-import static org.junit.Assert.*;
 
 public class SyncServiceTest {
 
